Depecrate any mcw-side path translations
I added translations (:IN: to a shm, etc) into the stdiowrap earlier
in order to support DOCK6 which requires "pseudo-filenames" to be
specified within a config file instead of on the command line.

Now that this exists, we don't need to translate with mcw when
building the commandline options for a search.  This removes some
memory allocations that caused issues on newton, and were the
cause of the recently discovered leak.

However, the translation in stdiowrap for :DB: was incorrect and is
now fixed.