Basic FASTA iterator
I call it basic, because I'd really like to be able to make the fasta sequence
"self-standing". The problem is, that BLAST won't know that the FASTA sequence
is done until it requests YET ANOTHER sequence just to receive the '>'
delimiter, or EOF.  It'd be better if we could just put a surrogate '>' at the
end of the sequence, or modify BLAST to handle an "end delimiter".